Revert "Improve execution name readability (#5637)"
This reverts commit 2ed24086f3d5eac5b779c1755642d4ac1b1d0f76.
