+++ flyteadmin/go.mod
@@ -48,7 +48,6 @@ require (
 	github.com/spf13/pflag v1.0.5
 	github.com/stretchr/testify v1.9.0
 	github.com/wI2L/jsondiff v0.5.0
-	github.com/wolfeidau/humanhash v1.1.0
 	go.opentelemetry.io/contrib/instrumentation/google.golang.org/grpc/otelgrpc v0.47.0
 	go.opentelemetry.io/otel v1.24.0
 	golang.org/x/oauth2 v0.16.0
