+++ flyteadmin/pkg/runtime/interfaces/application_configuration.go
@@ -49,8 +49,7 @@ type PostgresConfig struct {
 }
 
 type FeatureGates struct {
-	EnableArtifacts     bool `json:"enableArtifacts" pflag:",Enable artifacts feature."`
-	EnableFriendlyNames bool `json:"enableFriendlyNames" pflag:",Enable generation of friendly execution names feature."`
+	EnableArtifacts bool `json:"enableArtifacts" pflag:",Enable artifacts feature."`
 }
 
 // ApplicationConfig is the base configuration to start admin
